Fix review comments(add fund_address fn for sponsoring the main transaction), minor refactoring
